refactor: Move helpers + add jsdoc
